requirements.txt
openai
biopython
requests
beautifulsoup4
python-dotenv
pandas
thefuzz
